refactor(summaryContext): lift shared core into internal/summarycontext
The REST and MCP SummaryContext builders carried ~200 LoC of duplicated
implementation — the issue-index map, group-aware key arithmetic,
canonicalSingular kind normalization, BuildIssueIndex compose loop,
managedByFromRelationships extraction, and the per-row dispatch
closure body that picks namespaced vs cluster-wide index by scope. The
only substantive difference was the topology source (REST: broadcaster
cache; MCP: 5s memoizer).

Lift the shared core into a new internal/summarycontext package that
exports:
- Builder (the per-request closure type)
- IssueIndex with Count(group, kind, ns, name) int
- CanonicalSingular(kind) string
- BuildIssueIndex(provider, namespaces, kindFilter) IssueIndex
- ManagedByFromRelationships(rel) *ManagedByRef
- BuilderFromIndexes(topo, namespacedIdx, clusterIdx) Builder

BuilderFromIndexes takes the topology snapshot as an argument so REST
and MCP keep their respective sources unchanged — REST passes
s.broadcaster.GetCachedTopology(), MCP passes the memoized build.

internal/server/summary_context.go and internal/mcp/summary_context.go
shrink to thin wrappers that own their topology source and forward
everything else to the shared package. The MCP file additionally keeps
its local summaryCtxTopoMemo + buildSummaryContextTopology helpers
(MCP-specific, no analogue on the REST side).

Pure-function tests (issue-index key arithmetic, BuildIssueIndex over a
fake provider, CanonicalSingular, ManagedByFromRelationships, the
dual-index dispatch shape) move into the new package alongside the
shared code. The REST-side wiring tests (attachSummaryContextToList
end-to-end, issueIndexNamespaces dispatch) stay in internal/server
since they exercise handler plumbing. The MCP-side test file was
pure-function-only and is deleted outright — coverage now lives in
internal/summarycontext.

Net: ~390 LoC dropped from the tree; no wire-shape or behavior changes.
